Вынес общую логику обновления
